Include any required parents, not just the country
There was some logic in bin/build to make sure that the country was
present in the file, because otherwise it would be referenced as a
parent and it's good to have all ancestors of areas in the file so it's
useful standalone.  However, this was the wrong way to approach this,
really: sub-national regions can similarly need to be included to avoid
referencing missing parents. This commit changes the logic so that all
parents of areas that would be included are included.

I don't like this implementation, really; I think boundary_data.rb would
be clearer now if it had an area class rather than just handling hashes
with helper methods.