Handle Wikidata language codes with hyphens
Currently the values of the `config.json` `language_map` hash
are used directly to generate variable names in Wikidata
queries - this is problematic in the case of language codes
with hyphens as it generated syntactically invalid variable
names.

The current case that this is aimed at supporting is 'zh-tw',
which is slightly unusual in that it doesn't appear in
https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all
but is being used in labels (https://www.wikidata.org/wiki/Property_talk:P424).
However, there are other languages in the language code list that
also have hyphens, so I believe we will need this change even
if we make the decision only to support official Wikidata
language codes.