Add a 'seat_counts' attribute to legislatures
Some organizations are used both as executives and legislatures at the
moment, and we're going to merge the JSON objects for them together at
some point. When we do that, we need to preserve which position the seat
count actually refers to, so I've made seat_counts a hash which maps
position item IDs to the seat count.