Include the role_superclass for legislative positions as well
Consistently having a role_superclass turns out to be useful
downstream, so this commit adds it to the legislative data (it was
already present for executive data).

There is a slightly odd mismatch here, in that the role metadata that
the queries are built from is different between the legislative and
executive directories: for executive, the metadata specifies the
role_superclass, while for legislative it specifies the role.

It would be better if this were consistent, but I think it makes sense
to first get the data included in the build, and then change the
metadata and query generation so they're consistent, checking that the
data produced is the same.