Fix bug where missing seat count caused error
Includes tests to ensure a warning is generated when a seat count is missing
(instead of erroring), and to ensure that the seat count is exposed on
organization entities in the output.

Original work by Louise, and some subsequent wrangling by Alex.

Replace unused language_map param with wikidata_labels in order to
produce the warning message.