s/FILTER NOT EXISTS/MINUS/ in executive index query
This change makes the executive index query return far more quickly, resolving
the problem with timeouts documented in #45. Hence, this closes #45.

In almost all circumstances — including this one — `FILTER NOT EXISTS` and
`MINUS` are functionally equivalent, but `MINUS` is far more performant here.
In short, `FILTER NOT EXISTS` will be constructing the bracketed resultset anew
for each `?adminArea`, whereas `MINUS` will be constructing it once and
minusing those results from previous results.

No tests, as this is non-functional change.