Fix regression: only 'en' was appearing in label service template output
The regression was introduced in 9fb0986 when `languages` was removed
from the context and replaced with `config`, making `languages`
available as `config.languages`, without updating
`lib/commons/builder/queries/label_service.rq.liquid`.

Added a regression test, but we should also ensure that in general
attempting to access variables not in the context should raise an
exception.