Remove the inheritance dependency of WikidataQueries on WikidataClient
The primary concern of the WikidataQueries class is to generate SPARQL
queries based on Config; running the queries can be done by a
WikidataClient object without the need for this inheritance relationship
(which I think makes the code harder to understand, with no obvious
benefit). The inheritance relationship was also being used to make the
Config available to WikidataQueries in a hard-to-spot way.

This commit removes that inheritance relationship in favour of creating
WikidataClient objects for when a query needs to be performed, and
explicitly added config to the initialization of the object.