Reduce the complexity of Legislature.list
This was exceeding RuboCop's Metrics/AbcSize and Metrics/MethodLength
metrics, even at the increased values in .rubocop_todo.yml. This change
keeps their complexity under those limits, but it's not very pleasant -
those two methods have a lot of code in common.

This being a static method unfortunately makes it more difficult than
I'd like to simplify these in a more parisomious way, but this is
probably an improvement in readability for the moment.