Fetch rich position data for area-associated positions
In the downstream build, we need to be able to show for each area the
positions that are associated with it - not just the position item ID
(available from boundaries/build/index.json) but also what level it's at
(national, regional, local), what chamber or organization it's
associated with (e.g. a cabinet, a legislature), and whether it's an
executive or legislative position.

The results of us querying Wikidata for members of these positions will
bring in some of this data (and could be extended to the rest) but that
will only be for areas where we know who the representatives are. That's
not good enough - we should have a complete record of the areas we're
concerned with from the country from the data under boundaries/ and we
need to find this rich position data for each of those. (In fact, as
Alex Dutton pointed out in review, it's even more helpful to find this
data for all positions associated with any admin area in the country,
not just those already mentioned in the boundary data index.)

This commit adds an extra step to `update` and `build` to fetch and
parse this richer position data.

(Note that this will likely make some countries fail `build check`, but
that's a good thing - in those cases we need to update the boundary data
index so that this data *is* complete.)