Always use legislature position for exposed/serialized role.
If there's a term-specific position, we should use it to find results, but
should still expose the generic position ID in results, to hide the vagaries of
there being a term-specific position.

This also means that the associated positions in the boundaries index can use
the generic position ID, and won't need updating every time a new term starts
(with its associated position).

Closes #76.