Handle legislative positions that subclass other positions we return
If a house has multiple position 'has part's, and one subclasses
another, then we don't need to include the subclass position in the
legislative index. We can find holders of the subclass position in the
legislative membership query.

This doesn't handle the case of a house having 'has part' positions that
don't subclass each other, which will still cause duplicated entries in
the legislative index, for which we should warn.

Closes #87.