Deduplicate person links if multiple result rows
Previously, if a membership row was duplicated, we'd end up with a link
item in the person for every row, even if the link was the same. This
calls `.uniq!` on the links to ensure there are no duplicates.

This is tested by adding a duplicate row to the fixtures and checking
that it still produces the expected output data.