Ignore legislatures that have parts that are also legislatures.
This supports the case where a council organisation (e.g. "Bradford City
Council", Q16953331) is an instance of something (e.g. "English
metropolitan district council", Q19414242) subclasses legislature, but
has a separate part within it that is a legislature (e.g. "council of
Bradford City Council", Q55936924). In this case, we want the "council
of X" item, but to exclude the item it's part of.

The query change adds "legislature" alongside the existing case of
"legislative house" as "has part"s that preclude the inclusion of a
legislature.