Exclude memberships that overlap a term, but which are for another term
If a membership overlaps the term we're querying about through an
explicit start or end date, but it's also linked to another term then we
want to exclude it because it's likely not been updated with an end date
yet.

This is particularly important when querying about future terms, as
current memberships can't yet be given an end date, but it can't be
assumed that the membership also applies to the future term.

No tests as this is just a query change.